+++ src/components/App.js
@@ -2,7 +2,7 @@ import '../App.css';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Header from './Header';
 import HomePage from './Home';
-import WorksPage from './Works';
+import WorkPage from './Work';
 import DetailsPage from './DetailsPage';
 import AboutPage from './AboutPage';
 
@@ -20,8 +20,8 @@ const App = () => {
         <Routes>
           <Route path="/" element={<HomePage />} />
           <Route path="/about" element={<AboutPage />} />
-          <Route path="/works" element={<WorksPage />} />
-          <Route path="/works/:id" element={<DetailsPage />} />
+          <Route path="/projects" element={<WorkPage />} />
+          <Route path="/projects/:id" element={<DetailsPage />} />
         </Routes>
       </div>
     </BrowserRouter>
