+++ src/components/Project.js
@@ -41,7 +41,7 @@ const Project = (props) => {
   return (
     <li className="grid project w-full gap-5 lg:gap-4 2xl:gap-7 pt-5 px-2 2xl:px-16 pb-8 snap-start h-full overflow-y-scroll lg:overflow-y-auto animate-slide_in">
       <h2 className="title w-full text-center xl:text-left text-4xl 2xl:text-5xl 2xl:mt-10 text-primary-1000 dark:text-white font-medium capitalize">{title}</h2>
-      <Link to={`/works/${id}`} className="image-container mx-auto px-3 xl:px-1 xl:mt-5 2xl:mt-5 max-w-sm md:max-w-md lg:max-w-xs xl:max-w-none w-full xl:h-full xl:flex flex-col justify-center">
+      <Link to={`/projects/${id}`} className="image-container mx-auto px-3 xl:px-1 xl:mt-5 2xl:mt-5 max-w-sm md:max-w-md lg:max-w-xs xl:max-w-none w-full xl:h-full xl:flex flex-col justify-center">
         <Box className="w-full image">
           <img className="w-full" src={image} alt={title} />
         </Box>
@@ -77,7 +77,7 @@ const Project = (props) => {
           </IconButton>
           )}
         </Box>
-        <Link to={`/works/${id}`} className="xl:hidden text-white dark:text-primary-1000 bg-primary-1000 dark:bg-primary-50 flex gap-5 items-center py-1 px-3 uppercase font-semibold text-lg no-underline"> See More </Link>
+        <Link to={`/projects/${id}`} className="xl:hidden text-white dark:text-primary-1000 bg-primary-1000 dark:bg-primary-50 flex gap-5 items-center py-1 px-3 uppercase font-semibold text-lg no-underline"> See More </Link>
       </Box>
     </li>
   );
